Role::Tiny->make_role($package)
This method isolates the creation of a role
from the installation of subroutines (requires, with,
around, before, after) as done by Role::Tiny::import.

It may be used for programmatic creation of roles in runtime
(vs compile-time) and to make easier to extend and modify
the basic Role::Tiny features.