fixes to existing method detection for non-standard subs
There were various issues with detecting when stubs and constants were
changed, to detect if they should be treated as methods or not.

Previously, the code tried to avoid inflating constants when it
encountered them, in case they were stored directly in the stash as
subs, read only scalar refs, or similar.  Trying to properly record the
information about subs given all of the various things that can be
stored in the stash is very hard, and comparing them over time is
difficult as well.  Working around this would require a large amount of
very weird code, and all to save a minimal amount of memory.  Instead,
just use normal defined and exists checks.