fix constants being tracked as methods when imported before Role::Tiny but used after
After perl 5.10, constants will normally be created as scalar refs
in the symbol table.  If used as a coderef, they will automatically
inflate to a glob with a constant sub inside it.  But on perl 5.12 to
5.18, if the constant is then used as a constant after that, it will be
deflated back to a scalar ref.  If you then try to use it as a coderef,
a new coderef will be created.  This process can be repeated, creating
new coderefs each time.

In Role::Tiny, we need to track the identity of subs that exist before
importing Role::Tiny, because they are excluded from the list of methods
to compose, and if the methods are changed after importing, they should
not be excluded.  We use the identity of the coderef to do this, since
other options are more complex and less reliable.  Having perl
re-generate the coderefs breaks this tracking.  On the effected versions
of perl, we can populate the scalar slot of the glob.  This will prevent
perl from being able to convert the constant sub back into a scalar ref.
Since scalar slots in globs aren't detectable from pure perl if they
are undef, this seems like an acceptable compromise to work around the
behavior.