treat methods added via create_class_with_roles as direct methods, not inherited methods
create_class_with_roles uses pre-composed packages that can be added as
superclasses to achieve the same thing as applying the roles directly.
This makes creating multiple variants using the same roles less work.
But that means that the methods from those roles aren't copied into the
resulting package, so they wouldn't be taken into account when excluding
methods applied via a future apply_roles_to_package.

When installing methods to a package, we can use the list of applied
roles to build a full list of methods to exclude, not caring that the
methods don't exist directly in the package.