force properly working hash seed when testing against older Moo
The dependents test checks older versions of Moo.  Some older versions
have a test that is hash order sensitive and may fail on newer perls.
The dependents test tried to force a hash seed, but the seed needed can
depend on the perl version or configuration.

Implement a search to find a seed that gives the correct order for Moo's
test.