be more flexible about error to accept when testing applying a bad modifier
The code may be updated to rely on Class::Method::Modifiers to throw an
error, which will have a different form. We don't care what the specific
error message is, and our current error is confusing anyway. Adjust the
test to just check that the method name is included in the error.