use our own isa search rather than MRO::Compat when core function not available
If mro::get_linear_isa is not available, then we're on an old perl, and
MRO::Compat hasn't been loaded. This means the MRO can't have been
changed from DFS, which we can trivially implement ourselves rather than
rely on MRO::Compat. In fact, we don't care about the MRO at all for our
usage.

Still prefer the core function when it is available, since it should be
faster.

This means MRO::Compat is only needed to provide next::method, which is
used in create_class_with_roles.