fix applying all methods from roles after they participate in a conflict
If multiple roles were applied to a class, and they had a method
conflict, and that conflict was resolved by a provided method, the
methods would be permanently removed from the roles' method lists.

This was due to the code trying to localize entries in a number of
hashes by using postfix for, expecting it to not create a new scope.
Unfortunately, postfix for does create a scope, so the local had no
impact, and the delete immediately following it would be permanent.

To fix this, we have to use a really hack of using a goto loop, so that
the local is always in the same scope as the remainder of the function.
This can't be done with just a slice because we need to apply the local
to multiple hashes, not just multiple entries in a hash.