remove composable packages optimization
The composable packages were an optimization for roles used for
create_class_with_roles and apply_roles_to_object. "Composable" packages
would be generated that for any modifier, had a sub doing nothing but
calling next::method, and would have one role applied. These could then
be assembled using multiple inheritance to quickly construct a new
class. This avoids needing to do the full role application process for
every combination of roles.

However, this is a totally distinct code path, using an entirely
different process to try to match behavior. It was easy for the
implementations to drift from each other. It also has the problem of
using next::method. next::method always uses C3 method resolution, which
is an essential part of the required behavior for the composable
packages. But if the class is using multiple inheritance with the DFS
MRO, it will result in different behavior from manually creating a
subclass and applying roles. These methods will also be slower than a
normal modifier.

This doesn't seem like a worthwhile trade off. The composable package
behavior requires a significant amount of code, causes bugs, and makes
maintenance more difficult. The behavior it is optimizing for is also
a rather specialized case. It optimizes for creating new subclasses
using new combinations of the same set of roles. More likely, the same
set of classes and roles will be used, and that case is optimized for
separately.

Remove the composable package behavior, and add a test for the
difference in behavior of create_class_with_roles and manually creating
a subclass and applying the roles.