improve error for conflicts when creating a new class with roles
The error for method conflicts when applying roles using
create_class_with_roles should refer to the class being created from,
not the generated class. This was previously the case, but refactoring
to combine the code paths for various role application mechanisms
removed the separate error.

Store the class a generated class is based on, and use that base class
if it exists to generate the conflict errors.