disable dynamic_config and enable static install
We no longer have any dependencies that are conditional on the perl
version, so we can disable dynamic_config and enable x_static_install.

However, we were previously adding extra dependencies under
EXTENDED_TESTING, primarily for cpancover.com. We can't do that with
dynamic_config disabled, so we can't always run the extra tests under
EXTENDED_TESTING. Instead, we need to check if the developer
dependencies are available, and only then enable xt tests.

This will likely result in a lower score on cpancover.com. Hopefully
some time in the future it will be updated to install recommended
modules, which will allow the xt tests to be run.